[Server] - Move from main.dev.js to server.js
Allowing more moduler method handling and further refactoring in the future
